replace gorilla/mux route with http.ServeMux
